Fix 2Gb + add new service + fix bugs
Former-commit-id: cce70d24e3a0e8887285a666a5dea104de933196

+++ grails-app/domain/be/cytomine/image/UploadedFile.groovy
@@ -34,7 +34,7 @@ class UploadedFile extends CytomineDomain implements Serializable{
     String path
     String contentType
     UploadedFile parent
-    int size
+    Long size
     int status = 0
 
     /**
